Обновляет задание Отрисуй меня полностью (часть 2)
