+++ js/main.js
@@ -1,6 +1,7 @@
 import {showUsersPictures} from '/js/pictures.js';
-import {showFullImage} from '/js/full-image.js';
+import {watchClickThumbnail} from '/js/full-image.js';
+import {similarPosts} from '/js/utils.js';
 
-
-const posts = showUsersPictures();
-showFullImage(posts);
+const posts = similarPosts();
+showUsersPictures(posts);
+watchClickThumbnail(posts);
